requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
Faker==37.5.3
fonttools==4.58.4
gitdb==4.0.12
GitPython==3.1.44
holidays==0.75
idna==3.10
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.43.1
numpy==2.3.0
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.2.0
protobuf==6.31.1
pyarrow==20.0.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
six==1.17.0
smmap==5.0.2
streamlit==1.46.1
streamlit-aggrid==1.1.7
streamlit-calendar==1.3.1
streamlit-option-menu==0.4.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xlsxwriter==3.2.5
psycopg2-binary
streamlit-lottie
python-docx



